+++ Recurly/Resources/CouponRedemptionMini.cs
@@ -25,7 +25,7 @@ namespace Recurly.Resources
 
         /// <value>The amount that was discounted upon the application of the coupon, formatted with the currency.</value>
         [JsonProperty("discounted")]
-        public float? Discounted { get; set; }
+        public decimal? Discounted { get; set; }
 
         /// <value>Coupon Redemption ID</value>
         [JsonProperty("id")]
