+++ Recurly/Resources/PlanPricing.cs
@@ -21,11 +21,11 @@ namespace Recurly.Resources
 
         /// <value>Amount of one-time setup fee automatically charged at the beginning of a subscription billing cycle. For subscription plans with a trial, the setup fee will be charged at the time of signup. Setup fees do not increase with the quantity of a subscription plan.</value>
         [JsonProperty("setup_fee")]
-        public float? SetupFee { get; set; }
+        public decimal? SetupFee { get; set; }
 
         /// <value>Unit price</value>
         [JsonProperty("unit_amount")]
-        public float? UnitAmount { get; set; }
+        public decimal? UnitAmount { get; set; }
 
     }
 }
